# hotel-booking-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Booking API",
        "description": "REST API cho hệ thống đặt phòng khách sạn.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/admin/phong": {
            "get": {
                "tags": [
                    "Quản trị - Phòng"
                ],
                "summary": "Lấy danh sách phòng quản trị",
                "operationId": "e2370dbe9e7419cb0c3eec323e3b8c6a",
                "parameters": [
                    {
                        "name": "tim_kiem",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "maximum": 50,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy danh sách phòng thành công"
                    },
                    "401": {
                        "description": "Chưa xác thực"
                    },
                    "403": {
                        "description": "Không có quyền quản trị"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Quản trị - Phòng"
                ],
                "summary": "Thêm phòng kèm ảnh",
                "operationId": "e9a6fb1a9823f9be8537d186aedbc87a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "so_phong",
                                    "loai_phong_id",
                                    "gia_phong",
                                    "so_nguoi_toi_da"
                                ],
                                "properties": {
                                    "so_phong": {
                                        "type": "string"
                                    },
                                    "loai_phong_id": {
                                        "type": "integer"
                                    },
                                    "ten_phong": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "gia_phong": {
                                        "type": "number"
                                    },
                                    "so_nguoi_toi_da": {
                                        "type": "integer"
                                    },
                                    "mo_ta": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "hinh_anh": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Thêm phòng thành công"
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/phong/{id}": {
            "get": {
                "tags": [
                    "Quản trị - Phòng"
                ],
                "summary": "Lấy chi tiết phòng quản trị",
                "operationId": "7e1fe8cfe1d791730d5d8004aa2896a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy chi tiết phòng thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy phòng"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Quản trị - Phòng"
                ],
                "summary": "Cập nhật phòng kèm ảnh",
                "operationId": "f3ab7e9a6269981d4a4c8291d0b724a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "so_phong": {
                                        "type": "string"
                                    },
                                    "loai_phong_id": {
                                        "type": "integer"
                                    },
                                    "ten_phong": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "gia_phong": {
                                        "type": "number"
                                    },
                                    "so_nguoi_toi_da": {
                                        "type": "integer"
                                    },
                                    "mo_ta": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "hinh_anh": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật phòng thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy phòng"
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/phong/{id}/ngung-hoat-dong": {
            "put": {
                "tags": [
                    "Quản trị - Phòng"
                ],
                "summary": "Ngừng sử dụng phòng",
                "operationId": "0ceb8a023bb26a614969ce589ac93b1a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Đã ngừng sử dụng phòng"
                    },
                    "404": {
                        "description": "Không tìm thấy phòng"
                    },
                    "422": {
                        "description": "Phòng đã ngừng hoạt động"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/phong/{id}/kich-hoat": {
            "put": {
                "tags": [
                    "Quản trị - Phòng"
                ],
                "summary": "Kích hoạt lại phòng",
                "operationId": "3d2e1275b3fba8eaf9af34170d035b6f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Đã kích hoạt lại phòng"
                    },
                    "404": {
                        "description": "Không tìm thấy phòng"
                    },
                    "422": {
                        "description": "Phòng đang hoạt động"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/phong": {
            "get": {
                "tags": [
                    "Phòng"
                ],
                "summary": "Lấy danh sách phòng",
                "operationId": "c65fb6f9b8dab69da92e29c9fad163fc",
                "parameters": [
                    {
                        "name": "trang_thai",
                        "in": "query",
                        "description": "Lọc chính xác theo trạng thái phòng",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "loai_phong_id",
                        "in": "query",
                        "description": "Lọc chính xác theo ID loại phòng",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "so_nguoi",
                        "in": "query",
                        "description": "Lọc phòng có số người tối đa lớn hơn hoặc bằng giá trị này",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy dữ liệu thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lấy dữ liệu thành công"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Tham số truy vấn không hợp lệ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dữ liệu không hợp lệ"
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Bearer Token",
                "scheme": "bearer"
            }
        }
    },
...
}